Revert back into multiple components to fix layout on wide native devices.
