+++ src/pages/setpassword/index.native.js
@@ -1,0 +1,3 @@
+import SetPasswordPageNarrow from './SetPasswordPageNarrow';
+
+export default SetPasswordPageNarrow;
